+++ callautomation-quickstart/app.py
@@ -13,11 +13,22 @@ import json
 
 app = Flask(__name__)
 
+# Your ACS resource connection string
 ACS_CONNECTION_STRING = "<ACS_CONNECTION_STRING>"
+
+# Your ACS resource PSTN number will act as source number to start outbound call
 ACS_PSTN_SOURCE_NUMBER = "<ACS_PSTN_SOURCE_NUMBER>"
+
+# Target phone number you want to receive the call.
 TARGET_PHONE_NUMBER = "<TARGET_PHONE_NUMBER>"
-CALLBACK_EVENTS_URI = "<CALLBACK_HOST_WITH_PROTOCOL>/callback_events"
-WELCOME_MESSAGE_PROMPT = FileSource("WAV_FILE_URL")
+
+# Callback events URL to handle callback events, such as CallConnected.
+CALLBACK_HOST = "<CALLBACK_HOST_WITH_PROTOCOL>"
+CALLBACK_EVENTS_URL = CALLBACK_HOST + "/callback_events"
+
+# These recorded prompts must be uploaded to publicly available URLs
+MAIN_MENU_PROMPT_URL = "<MAIN_MENU_WAV_FILE_URL>"
+
 CALL_CONNECTION_ID = None
 CALLBACK_EVENTS = []
 
@@ -49,10 +60,10 @@ def create_call_handler():
     target_participant = PhoneNumberIdentifier(TARGET_PHONE_NUMBER)
     source_caller_id_number = PhoneNumberIdentifier(ACS_PSTN_SOURCE_NUMBER)
     call_invite = CallInvite(target=target_participant, source_caller_id_number=source_caller_id_number)
-    call_automation = CallAutomationClient.from_connection_string(ACS_CONNECTION_STRING)
-    create_call_result = call_automation.create_call(call_invite, CALLBACK_EVENTS_URI)
+    call_automation_client = CallAutomationClient.from_connection_string(ACS_CONNECTION_STRING)
+    call_connection_properties = call_automation_client.create_call(call_invite, CALLBACK_EVENTS_URL)
     global MENU_TEMPLATE_STR, CALL_CONNECTION_ID
-    CALL_CONNECTION_ID = create_call_result.call_connection_id
+    CALL_CONNECTION_ID = call_connection_properties.call_connection_id
     MENU_TEMPLATE_STR = MENU_TEMPLATE_STR.replace("<CALL_CONNECTION_ID>", CALL_CONNECTION_ID)
     return redirect("/")
 
@@ -60,18 +71,20 @@ def create_call_handler():
 @app.route('/record_call')
 def record_call():
     call_connection_id = request.args.get('call_connection_id')
-    server_call_id = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id).get_call_properties().server_call_id
-    call_automation = CallAutomationClient.from_connection_string(ACS_CONNECTION_STRING)
-    call_automation.start_recording(call_locator=ServerCallLocator(server_call_id),
-                                    recording_state_callback_url=CALLBACK_EVENTS_URI)
+    server_call_id = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING,
+                                                                 call_connection_id).get_call_properties().server_call_id
+    call_automation_client = CallAutomationClient.from_connection_string(ACS_CONNECTION_STRING)
+    call_automation_client.start_recording(call_locator=ServerCallLocator(server_call_id),
+                                           recording_state_callback_url=CALLBACK_EVENTS_URL)
     return redirect("/")
 
 
 @app.route('/play_audio')
 def play_audio_handler():
     call_connection_id = request.args.get('call_connection_id')
-    call_connection = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
-    call_connection.play_media_to_all(WELCOME_MESSAGE_PROMPT)
+    call_connection_client = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
+    file_source = FileSource(MAIN_MENU_PROMPT_URL)
+    call_connection_client.play_media_to_all(file_source)
     return redirect("/")
 
 
@@ -79,23 +92,24 @@ def play_audio_handler():
 def recognize_dtmf_handler():
     call_connection_id = request.args.get('call_connection_id')
     target_participant = PhoneNumberIdentifier(TARGET_PHONE_NUMBER)
-    call_connection = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
-    call_connection.start_recognizing_media(input_type=RecognizeInputType.DTMF,
-                                            target_participant=target_participant,
-                                            play_prompt=WELCOME_MESSAGE_PROMPT,
-                                            interrupt_prompt=True,
-                                            initial_silence_timeout=10,
-                                            dtmf_inter_tone_timeout=10,
-                                            dtmf_max_tones_to_collect=1,
-                                            dtmf_stop_tones=[DtmfTone.POUND, DtmfTone.ASTERISK])
+    file_source = FileSource(MAIN_MENU_PROMPT_URL)
+    call_connection_client = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
+    call_connection_client.start_recognizing_media(input_type=RecognizeInputType.DTMF,
+                                                   target_participant=target_participant,
+                                                   play_prompt=file_source,
+                                                   interrupt_prompt=True,
+                                                   initial_silence_timeout=10,
+                                                   dtmf_inter_tone_timeout=10,
+                                                   dtmf_max_tones_to_collect=1,
+                                                   dtmf_stop_tones=[DtmfTone.POUND])
     return redirect("/")
 
 
 @app.route('/hangup_call')
 def hangup_call_handler():
     call_connection_id = request.args.get('call_connection_id')
-    call_connection = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
-    call_connection.hang_up(is_for_everyone=True)
+    call_connection_client = CallConnectionClient.from_connection_string(ACS_CONNECTION_STRING, call_connection_id)
+    call_connection_client.hang_up(is_for_everyone=True)
     global MENU_TEMPLATE_STR
     MENU_TEMPLATE_STR = MENU_TEMPLATE_STR.replace(CALL_CONNECTION_ID, "<CALL_CONNECTION_ID>")
     return redirect("/")
